listings done, login/signup done, creating/updating/deleting listings to assositated users.

+++ backend/src/serviceListings/listing.service.ts
@@ -1,7 +1,7 @@
-import { Injectable, Inject, NotFoundException } from '@nestjs/common';
+import { Injectable, Inject, NotFoundException, ForbiddenException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Listing } from './listing.entity';
-import { User } from '../users/user.entity'; // Adjust the path as necessary
+import { User } from '../users/user.entity';
 
 @Injectable()
 export class ListingService {
@@ -15,7 +15,7 @@ export class ListingService {
   }
 
   async findOne(id: number): Promise<Listing> {
-    const listing = await this.listingRepository.findOne({ where: { id } });
+    const listing = await this.listingRepository.findOne({ where: { id }, relations: ['user'] });
     if (!listing) {
       throw new NotFoundException(`Listing with ID ${id} not found`);
     }
@@ -23,17 +23,33 @@ export class ListingService {
   }
 
   async create(user: User, listingData: Partial<Listing>): Promise<Listing> {
-    const listing = this.listingRepository.create({ ...listingData, user });
+    const listing = this.listingRepository.create({
+      ...listingData,
+      user: user,
+    });
+    // console.log('this is what is logged:', listing);
     return await this.listingRepository.save(listing);
   }
 
-  async update(id: number, updateData: Partial<Listing>): Promise<Listing> {
-    await this.findOne(id);
+  async update(user: User, id: number, updateData: Partial<Listing>): Promise<Listing> {
+    const listing = await this.findOne(id);
+    
+    if (listing.user.id !== user.id) {
+      throw new ForbiddenException('You are not allowed to update this listing');
+    }
+  
     await this.listingRepository.update(id, updateData);
+  
     return this.findOne(id);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(user: User, id: number): Promise<void> {
+    const listing = await this.findOne(id);
+    console.log("this is remove", listing)
+    if (listing.user.id !== user.id) {
+      throw new ForbiddenException('You are not allowed to delete this listing');
+    }
+    
     const result = await this.listingRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Listing with ID ${id} not found`);
